added sample blog archive page
Archive pages are currently the same format as blog pages and are
essentially just recreations of the blog page at the time of archival.
When a blog post "expires" a new archival page (or pages) will be
created to reference these posts. These archival pages could either
be locked from further edits or open to customization based on the
needs of the site.