Add type~html
Likely the name should change to `text`
or `prose` or somesuch...right now it
matches `widget` for all intents and purposes.

In the old BlueInk the `html` type was a
TinyMCE enhanced editing space for the
typical content of a text-focused site.

It was also referred to as `text` in the UI.

So...